Test: Fix Wrong Return Type
See: https://mantis.ilias.de/view.php?id=35781

+++ Modules/TestQuestionPool/classes/class.assQuestion.php
@@ -2719,7 +2719,7 @@ abstract class assQuestion
      */
     abstract public function calculateReachedPoints($active_id, $pass = null, $authorizedSolution = true, $returndetails = false);
 
-    public function deductHintPointsFromReachedPoints(ilAssQuestionPreviewSession $previewSession, $reachedPoints): ?int
+    public function deductHintPointsFromReachedPoints(ilAssQuestionPreviewSession $previewSession, $reachedPoints): ?float
     {
         global $DIC;
